move e2e tests
